Renamed feed retreival and parsing tool.
