+++ functions/tools.py
@@ -2,30 +2,33 @@
 
 import json
 import logging
-import functions.helper_functions as helper_funcs
+import functions.feed_extraction as extraction_funcs
 
 
-def get_content(website: str) -> list:
-    '''Gets RSS feed content from a given website.
+def get_feed(website: str) -> list:
+    '''Gets RSS feed content from a given website. Can take a website or RSS
+    feed URL directly, or the name of a website. Will attempt to find RSS
+    feed and return title, summary and link to full article for most recent
+    items in feed
     
     Args:
-        website_url: URL or nam of website to extract RSS feed content from
+        website: URL or name of website to extract RSS feed content from
 
     Returns:
-        List of titles for the 10 most recent entries in the RSS feed from the
-        requested website.
+        JSON string containing the feed content or 'No feed found' if a RSS
+        feed for the requested website could not be found
     '''
 
     logger = logging.getLogger(__name__ + '.get_content')
     logger.info('Getting feed content for: %s', website)
 
-    feed_uri = helper_funcs.find_feed_uri(website)
+    feed_uri = extraction_funcs.find_feed_uri(website)
     logger.info('find_feed_uri() returned %s', feed_uri)
 
     if 'No feed found' in feed_uri:
         return 'No feed found'
 
-    content = helper_funcs.parse_feed(feed_uri)
+    content = extraction_funcs.parse_feed(feed_uri)
     logger.info('parse_feed() returned %s entries', len(list(content.keys())))
 
     return json.dumps(content)
